pyproject.toml
[project]
name = "portfolio-app2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.3.0 ; python_full_version >= '3.9'",
    "google-generativeai>=0.8.5",
    "google-generativeai>=0.7.0 ; python_full_version >= '3.9'",
    "numpy>=1.24.0",
    "openai>=1.102.0",
    "requests>=2.31.0",
    "scikit-learn>=1.3.0",
    "streamlit>=1.36.0",
    "streamlit-option-menu>=0.3.6",
]
